refactor: CrossMoveBehavior -> RotateMoveBehavior 의도 명확화 변경
