+++ src/main/java/janggi/domain/piece/behavior/rotatemove/Horse.java
@@ -1,4 +1,4 @@
-package janggi.domain.piece.behavior.crossmove;
+package janggi.domain.piece.behavior.rotatemove;
 
 import janggi.domain.Board;
 import janggi.domain.Side;
@@ -8,7 +8,7 @@ import janggi.domain.move.Vectors;
 import java.util.List;
 import java.util.Set;
 
-public final class Horse extends CrossMoveBehavior {
+public final class Horse extends RotateMoveBehavior {
 
     @Override
     protected List<Vectors> getVectorsList() {
