+++ src/main/java/janggi/factory/PieceFactory.java
@@ -7,8 +7,8 @@ import janggi.domain.piece.PieceBehavior;
 import janggi.domain.piece.behavior.General;
 import janggi.domain.piece.behavior.Guard;
 import janggi.domain.piece.behavior.Soldier;
-import janggi.domain.piece.behavior.crossmove.Elephant;
-import janggi.domain.piece.behavior.crossmove.Horse;
+import janggi.domain.piece.behavior.rotatemove.Elephant;
+import janggi.domain.piece.behavior.rotatemove.Horse;
 import janggi.domain.piece.behavior.straightmove.Cannon;
 import janggi.domain.piece.behavior.straightmove.Chariot;
 import java.util.HashMap;
